Split marketing content surfaces
