+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-prep-links.ts
@@ -1,0 +1,97 @@
+import type { AppLocale } from '@/i18n/locales';
+import type { PrepLinksSection } from '../_components/ModelPrepLinksSection';
+
+const NANO_BANANA_MODEL_SLUGS = new Set([
+  'nano-banana',
+  'nano-banana-pro',
+  'nano-banana-2',
+  'gpt-image-2',
+]);
+
+const VIDEO_PREP_MODEL_SLUGS = new Set([
+  'veo-3-1',
+  'kling-3-pro',
+  'happy-horse-1-0',
+  'sora-2-pro',
+  'ltx-2-3-pro',
+  'ltx-2-3-fast',
+]);
+
+export function buildModelPrepLinksSection(modelSlug: string, locale: AppLocale): PrepLinksSection | null {
+  const isNanoBananaFamily = NANO_BANANA_MODEL_SLUGS.has(modelSlug);
+  const isVideoPrepModel = VIDEO_PREP_MODEL_SLUGS.has(modelSlug);
+
+  if (!isNanoBananaFamily && !isVideoPrepModel) {
+    return null;
+  }
+
+  if (locale === 'fr') {
+    return isNanoBananaFamily
+      ? {
+          eyebrow: 'Avant de générer',
+          title: 'Préparez la référence avant l’edit',
+          body: 'Si l’image a d’abord besoin d’une référence personnage réutilisable ou d’un meilleur angle, réglez ça avant Nano Banana.',
+          links: [
+            { href: '/tools/character-builder', label: 'Créer une référence personnage réutilisable' },
+            { href: '/tools/angle', label: "Changer le point de vue avant l'edit" },
+            { href: '/app/image', label: 'Ouvrir Image' },
+          ],
+        }
+      : {
+          eyebrow: 'Avant de générer',
+          title: 'Préparez le frame avant la vidéo',
+          body: 'Verrouillez le personnage, corrigez l’angle ou construisez l’image source avant de dépenser des crédits en motion.',
+          links: [
+            { href: '/tools/character-builder', label: 'Conserver le même personnage' },
+            { href: '/tools/angle', label: 'Changer le point de vue avant la vidéo' },
+            { href: '/app/image', label: "Construire l'image source dans Image" },
+          ],
+        };
+  }
+
+  if (locale === 'es') {
+    return isNanoBananaFamily
+      ? {
+          eyebrow: 'Antes de generar',
+          title: 'Prepara la referencia antes del edit',
+          body: 'Si la imagen primero necesita una referencia de personaje reutilizable o un mejor ángulo, resuélvelo antes de Nano Banana.',
+          links: [
+            { href: '/tools/character-builder', label: 'Crear una referencia de personaje reutilizable' },
+            { href: '/tools/angle', label: 'Cambiar el punto de vista antes del edit' },
+            { href: '/app/image', label: 'Abrir Image' },
+          ],
+        }
+      : {
+          eyebrow: 'Antes de generar',
+          title: 'Prepara el frame antes del video',
+          body: 'Fija el personaje, corrige el ángulo o construye la imagen base antes de gastar créditos en motion.',
+          links: [
+            { href: '/tools/character-builder', label: 'Mantener el mismo personaje' },
+            { href: '/tools/angle', label: 'Cambiar el punto de vista antes del video' },
+            { href: '/app/image', label: 'Construir la imagen base en Image' },
+          ],
+        };
+  }
+
+  return isNanoBananaFamily
+    ? {
+        eyebrow: 'Before you generate',
+        title: 'Build the reference before the edit',
+        body: 'If the still needs a reusable character reference or a better viewpoint first, solve that before Nano Banana.',
+        links: [
+          { href: '/tools/character-builder', label: 'Build a reusable character reference' },
+          { href: '/tools/angle', label: 'Change the viewpoint before the edit' },
+          { href: '/app/image', label: 'Open Image' },
+        ],
+      }
+    : {
+        eyebrow: 'Before you generate',
+        title: 'Prepare the frame before video',
+        body: 'Lock the character, fix the viewpoint, or build the source still before you spend credits on motion.',
+        links: [
+          { href: '/tools/character-builder', label: 'Keep the character consistent' },
+          { href: '/tools/angle', label: 'Change the viewpoint before video' },
+          { href: '/app/image', label: 'Build the source still in Image' },
+        ],
+      };
+}
